AlbumManager Implementation
Seperated album specific functions to a seperate class that was inside the ScreenshotHandler
